# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swat Flood Early-Warning System API",
    "description": "MVP demo backend - temporary implementation for university evaluation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reports/": {
      "post": {
        "tags": [
          "community-reports"
        ],
        "summary": "Submit Report",
        "description": "Submit a community report.\n\nAccepts either zone_id or lat/lon coordinates.\nIf coordinates provided, finds nearest zone.\n\nTODO: Add auth, move to use-cases layer, add proper error handling.",
        "operationId": "submit_report_api_reports__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityReportCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "community-reports"
        ],
        "summary": "Get Reports",
        "description": "Get community reports with composable optional filtering.\n\nFilters compose: ``status=approved&zone_id=X`` returns approved reports in X.\nNo filters returns all reports (unbounded \u2014 caller should paginate in future).\n\nUsed for moderation list and official view.\n\nTODO: Add auth, move to use-cases layer, add proper error handling, add pagination.",
        "operationId": "get_reports_api_reports__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "zone_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zone Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports": {
      "post": {
        "tags": [
          "community-reports"
        ],
        "summary": "Submit Report",
        "description": "Submit a community report.\n\nAccepts either zone_id or lat/lon coordinates.\nIf coordinates provided, finds nearest zone.\n\nTODO: Add auth, move to use-cases layer, add proper error handling.",
        "operationId": "submit_report_api_reports_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityReportCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "community-reports"
        ],
        "summary": "Get Reports",
        "description": "Get community reports with composable optional filtering.\n\nFilters compose: ``status=approved&zone_id=X`` returns approved reports in X.\nNo filters returns all reports (unbounded \u2014 caller should paginate in future).\n\nUsed for moderation list and official view.\n\nTODO: Add auth, move to use-cases layer, add proper error handling, add pagination.",
        "operationId": "get_reports_api_reports_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "zone_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zone Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}": {
      "patch": {
        "tags": [
          "community-reports"
        ],
        "summary": "Update Report Status",
        "description": "Approve or reject a pending report.\n\nRequires admin JWT.",
        "operationId": "update_report_status_api_reports__report_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityReportUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommunityReportCreateRequest": {
        "properties": {
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "description": "Type of report: flooding, blocked_drain, high_river_level, etc."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Detailed description of the issue"
          },
          "reporter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Name",
            "description": "Optional reporter name"
          },
          "reporter_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Phone",
            "description": "Optional reporter phone number"
          },
          "zone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id",
            "description": "Zone ID (if known)"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude",
            "description": "Latitude coordinate (if zone_id not provided)"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude",
            "description": "Longitude coordinate (if zone_id not provided)"
          }
        },
        "type": "object",
        "required": [
          "report_type",
          "description"
        ],
        "title": "CommunityReportCreateRequest",
        "description": "Request schema for submitting a community report.",
        "example": {
          "description": "Water level rising rapidly near bridge",
          "latitude": 35.1234,
          "longitude": 72.5678,
          "report_type": "flooding",
          "reporter_name": "John Doe",
          "reporter_phone": "+1234567890"
        }
      },
      "CommunityReportListResponse": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/CommunityReportResponse"
            },
            "type": "array",
            "title": "Reports"
          },
          "reports_by_zone": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "$ref": "#/components/schemas/CommunityReportResponse"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reports By Zone"
          }
        },
        "type": "object",
        "required": [
          "reports"
        ],
        "title": "CommunityReportListResponse",
        "description": "Response schema for list of community reports."
      },
      "CommunityReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "reporter_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Name"
          },
          "reporter_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Phone"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "admin_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "zone_id",
          "report_type",
          "description",
          "submitted_at",
          "status"
        ],
        "title": "CommunityReportResponse",
        "description": "Response schema for a community report."
      },
      "CommunityReportUpdateRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "New status: Approved or Rejected"
          },
          "admin_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes",
            "description": "Optional admin notes"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CommunityReportUpdateRequest",
        "description": "Request schema for updating a report status.",
        "example": {
          "admin_notes": "Verified with local authorities",
          "status": "Approved"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}